[Adaptive] [Canonical Layouts] Added hero element to small screen configurations.
The hero element was missing from compact width size screens in our single view hero canonical layout demo. This was WAI/following the mocks, but it turns out this was a "design bug" and the mocks were wrong. The demo is now updated to include the hero element as the point of focus.

Also updated the title text appearance and a whole bunch of margins that weren't to spec or could use improvement.

PiperOrigin-RevId: 429571537