[Chips] [Docs] Fixed image description text.
The number of chips was mismatched with the image. Removed the number from the image description (but kept it in the a11y content description) because it's not relevant to the general reader -- but it is helpful for a11y purposes. Getting rid of the numbers makes for future easier maintenance.

PiperOrigin-RevId: 432218802