[Catalog][BottomSheet] Fix full-screen dialog layout after restrict expansion is enabled once
In the original implementation, we wrongly applied layout height on the bottom sheet dialog root view, instead of its content, which makes full-screen dialog broken since it's setting layout height to the content instead. Refactors the logic so the layout height will always be set to the content, instead of the root view.

PiperOrigin-RevId: 450932326