[FAB] Fix the issue that state-list animator cannot be customized
Before this CL we always reset the state-list animator when FAB elevation changes, i.e., any state-list animators set via XML will be overridden because elevation is set later. Fixes this by storing the default state-list animator created in onElevationsChanged() and check if it's been overridden before we update and reset it.

Resolves https://github.com/material-components/material-components-android/issues/1218

PiperOrigin-RevId: 452073249