[Catalog] Add ability to favorite a specific demo fragment/activity to launch on start
Heart icons are available in each demo row on the demo landing pages. Selecting one will make that demo launch immediately when opening the Catalog from a cold start. In order to make sure the back stack is correct when a specific demo is favorited, on launch the DemoLandingFragment for the favorited demo will be started first, and then the favorited demo will be started on top of that.

PiperOrigin-RevId: 461619522