[BottomSheetBehavior] Add internal method to disable hiding gestures
Add an internal method to disable hiding gestures, even when isHideable
is enabled.

Some users need to enable hideable so that they can set a state
programmatically, but still want to disable hiding in the UI, as the
hiding action would represent something too destructive.

This override preserves the beheavior of isHideable, while still
allowing a trapdoor to disable hiding gestures for whose who need to.