[Catalog] Added search/filter functionality to MDC Catalog.
This speeds up our development process and enhances the user experience of our catalog.

Also added a bottom line to the last row of boxes in the table of contents grid in order to avoid this visual breakage when the grid does not fill up the screen: screenm/9V8HPkufqM9BkVc.png. Previously, GridDividerDecoration avoided drawing the bottom line of the grid in the last row of the grid, but now that there can be less boxes than the screen size it's important to handle that case.

PiperOrigin-RevId: 465643774