[BottomSheet] Sync custom actions with drag handle views
Custom actions need to be set directly on the focused child views to make talkback announce the existence of those actions correctly, despite that when you open custom action menu you can actually see they are being inherited from the parent view.

Makes BottomSheetBehavior be aware of the existence of accessibility delegate views, and update the custom actions on it when needed.

PiperOrigin-RevId: 478804858