Draw label directly to canvas if label behavior is LABEL_VISIBLE
This allows the label to scroll with the Slider as expected. There's no need for the label to be drawn over all content if it's always going to be shown. In that case, it can be drawn directly to canvas since there will be adequate padding to ensure it's not overlapping adjacent views anyway.
Remove intrinsic padding from the Tooltip used for the label. Open to feedback here if this is the best solution, but was running into https://github.com/material-components/material-components-android/issues/2942 at the left-most bound while drawing on the canvas.