[ProgressIndicator] Fix linear progress indicator with corner radius
When corner radius of a linear progress indicator is larger than 0, the behavior it's confusing because when the progress fraction is slightly larger than 0, the progress indicator will suddenly show with a width of twice the corner radius.

Fixes this by adjusting the indicator startX and endX correctly so the visible indicator length is always proportional to the progress.

Resolves https://github.com/material-components/material-components-android/issues/2940

PiperOrigin-RevId: 509552399