Migrate `annotation-experimental` to `aar_import` and remove old lint target
Fixes failures in `material3` and `tiles` by opting in.

Fixes failures in `media3` and `protolayout` by disabling the check. This is not ideal, since it does not enable cross-language checks, but it is no worse than it was before.

Adds explicit `kotlin-stdlib` dependency where it was previously implicitly included by a poorly-configured `annotation-experimental` target.

Removes explicit `proguard.txt` reference from Car build, since it is already included in the AAR.

PUBLIC: Handle calls to @RequiresOptIn-annotated APIs
PiperOrigin-RevId: 584657528