[FloatingToolbar] Added materialThemeOverlays to automatically set orientation and the proper icon button style on floating toolbar children.
These new materialThemeOverlays allow the FloatingToolbar style variants to elegantly handle configuration of its children, by automatically setting the orientation of the .Horizontal style variants' children to horizontal and the orientation of the .Vertical variants' children to vertical, and setting the appropriate icon button style for each floating toolbar variant, as determined by the parent style.

Also added vibrant icon button styles so that they can be set from the new theme overlays, so that the proper button styling is set on both checkable and non-checkable buttons on both vibrant and non-vibrant variants.

Reconfigured the catalog demo to remove android:orientation from the children since it's no longer necessary.

PiperOrigin-RevId: 699262068