[TextField] Fix the issue that menu items gets filtered after recreation
It's actually a bug of framework's AutoCompleteTextView. When restoring the
state, it always filters the item list regardless the previously input item
is from menu selection or manual input.

This change fixes the issue by implementing the custom logic of input text
saving/restoring, and properly remember if the last input text filters the
item list or not.

Resolves https://github.com/material-components/material-components-android/issues/1464

PiperOrigin-RevId: 716328564