[OverflowLinearLayout] Added OverflowLinearLayout to support adaptive behavior of the floating and docked toolbars.
When wrapping the floating/docked toolbar's children in a OverflowLinearLayout, the toolbar will automatically move children to an overflow menu when there's not enough screen space for all the items to fit in the toolbar.

The children moved to an overflow menu will always be the ones at the end of the toolbar that do not fit the max toolbar size. An overflow button will be automatically placed as the last child of the toolbar when overflow happens.

PiperOrigin-RevId: 757750540