[TopAppBar][A11y] Improve keyboard navigation in TopAppBar catalog demos
The NestedScrollView in the Top App Bar demos would incorrectly receive keyboard focus. While this allowed for keyboard scrolling, the scroll container itself is not an interactive element and shouldn't be a focus stop.

This change fixes the issue by making the content within the scroll view focusable. The text content has been split into multiple, clickable TextViews. This makes the TextViews interactive elements, allowing them to correctly receive keyboard focus and improving the accessibility of the demos.

The logic for this has been extracted into a DemoUtils.setupClickableContentText method and applied across the relevant Top App Bar and Side Sheet demos.

PiperOrigin-RevId: 804881130